ipacm: fix using right size for L2TP session_id.
L2TP session is of 4 bytes. Make changes to use right size otherwise
L2TP data transfer fails.

+++ ipacm/inc/IPACM_Defs.h
@@ -455,7 +455,7 @@ struct l2tp_vlan_mapping_info
 {
 	/* the following are l2tp iface info (name, session id) */
 	char l2tp_iface_name[IPA_RESOURCE_NAME_MAX];
-	uint8_t l2tp_session_id;
+	uint32_t l2tp_session_id;
 	/* Add support for L2TP over UDP. */
 #ifdef IPA_L2TP_TUNNEL_UDP
 	enum ipa_l2tp_tunnel_type tunnel_type;
